+++ src/pages/TodayHabitPage/TodayHabitPage.jsx
@@ -1,16 +1,16 @@
 import React, { useEffect, useState } from 'react';
 import styles from './TodayHabitPage.module.css';
-import HabitList from '../../components/HabitComponents/HabitList';
-import HabitListHeader from '../../components/HabitComponents/HabitListHeader';
 import { useParams } from 'react-router-dom';
-import HabitConfirmModal from '../../components/HabitComponents/HabitConfirmModal';
-import CurrentTime from '../../components/CurrentTime/CurrentTime';
-import HabitHeader from '../../components/HabitComponents/HabitHeader';
 import {
   getTodayHabits,
   postHabit,
   toggleHabit,
 } from '../../services/HabitService';
+import HabitHeader from './HabitComponents/HabitHeader';
+import HabitConfirmModal from './HabitComponents/HabitConfirmModal';
+import HabitList from './HabitComponents/HabitList';
+import HabitListHeader from './HabitComponents/HabitListHeader';
+import CurrentTime from '../../components/CurrentTime/CurrentTime';
 
 const TodayHabitPage = () => {
   const [studyName, setStudyName] = useState('');
